Move all tests to individual suites
