+++ packages/react/src/checkbox/root/CheckboxRoot.test.tsx
@@ -515,6 +515,56 @@ describe('<Checkbox.Root />', () => {
 
       expect(button).not.to.have.attribute('data-focused');
     });
+
+    it('prop: validate', async () => {
+      await render(
+        <Field.Root validate={() => 'error'}>
+          <Checkbox.Root data-testid="button" />
+          <Field.Error data-testid="error" />
+        </Field.Root>,
+      );
+
+      const button = screen.getByTestId('button');
+
+      expect(button).not.to.have.attribute('aria-invalid');
+
+      fireEvent.focus(button);
+      fireEvent.blur(button);
+
+      expect(button).to.have.attribute('aria-invalid', 'true');
+    });
+
+    it('Field.Label', async () => {
+      const { container } = await render(
+        <Field.Root>
+          <Checkbox.Root data-testid="button" />
+          <Field.Description data-testid="description" />
+        </Field.Root>,
+      );
+
+      const internalInput = container.querySelector<HTMLInputElement>('input[type="checkbox"]');
+
+      expect(internalInput).to.have.attribute(
+        'aria-describedby',
+        screen.getByTestId('description').id,
+      );
+    });
+
+    it('Field.Description', async () => {
+      const { container } = await render(
+        <Field.Root>
+          <Checkbox.Root data-testid="button" />
+          <Field.Description data-testid="description" />
+        </Field.Root>,
+      );
+
+      const internalInput = container.querySelector<HTMLInputElement>('input[type="checkbox"]');
+
+      expect(internalInput).to.have.attribute(
+        'aria-describedby',
+        screen.getByTestId('description').id,
+      );
+    });
   });
 
   it('should change state when clicking the checkbox if it has a wrapping label', async () => {
