+++ packages/react/src/field/root/FieldRoot.test.tsx
@@ -179,131 +179,6 @@ describe('<Field.Root />', () => {
 
       expect(control).to.have.attribute('aria-invalid', 'true');
     });
-
-    describe('component integration', () => {
-      it('supports Checkbox', async () => {
-        await render(
-          <Field.Root validate={() => 'error'}>
-            <Checkbox.Root data-testid="button" />
-            <Field.Error data-testid="error" />
-          </Field.Root>,
-        );
-
-        const button = screen.getByTestId('button');
-
-        expect(button).not.to.have.attribute('aria-invalid');
-
-        fireEvent.focus(button);
-        fireEvent.blur(button);
-
-        expect(button).to.have.attribute('aria-invalid', 'true');
-      });
-
-      it('supports Switch', async () => {
-        await render(
-          <Field.Root validate={() => 'error'}>
-            <Switch.Root data-testid="button" />
-            <Field.Error data-testid="error" />
-          </Field.Root>,
-        );
-
-        const button = screen.getByTestId('button');
-
-        expect(button).not.to.have.attribute('aria-invalid');
-
-        fireEvent.focus(button);
-        fireEvent.blur(button);
-
-        expect(button).to.have.attribute('aria-invalid', 'true');
-      });
-
-      it('supports NumberField', async () => {
-        await render(
-          <Field.Root validate={() => 'error'}>
-            <NumberField.Root>
-              <NumberField.Input />
-            </NumberField.Root>
-            <Field.Error data-testid="error" />
-          </Field.Root>,
-        );
-
-        const input = screen.getByRole('textbox');
-
-        expect(input).not.to.have.attribute('aria-invalid');
-
-        fireEvent.focus(input);
-        fireEvent.blur(input);
-
-        expect(input).to.have.attribute('aria-invalid', 'true');
-      });
-
-      it('supports Slider', async () => {
-        const { container } = await render(
-          <Field.Root validate={() => 'error'}>
-            <Slider.Root>
-              <Slider.Control>
-                <Slider.Thumb data-testid="thumb" />
-              </Slider.Control>
-            </Slider.Root>
-            <Field.Error data-testid="error" />
-          </Field.Root>,
-        );
-
-        // eslint-disable-next-line testing-library/no-node-access
-        const input = container.querySelector<HTMLInputElement>('input')!;
-        const thumb = screen.getByTestId('thumb');
-
-        expect(input).not.to.have.attribute('aria-invalid');
-
-        fireEvent.focus(thumb);
-        fireEvent.blur(thumb);
-
-        expect(input).to.have.attribute('aria-invalid', 'true');
-      });
-
-      it('supports RadioGroup', async () => {
-        await render(
-          <Field.Root validate={() => 'error'}>
-            <RadioGroup data-testid="group">
-              <Radio.Root value="1">One</Radio.Root>
-              <Radio.Root value="2">Two</Radio.Root>
-            </RadioGroup>
-            <Field.Error data-testid="error" />
-          </Field.Root>,
-        );
-
-        const group = screen.getByTestId('group');
-
-        expect(group).not.to.have.attribute('aria-invalid');
-
-        fireEvent.focus(group);
-        fireEvent.blur(group);
-
-        expect(group).to.have.attribute('aria-invalid', 'true');
-      });
-
-      it('supports Select', async () => {
-        await render(
-          <Field.Root validate={() => 'error'}>
-            <Select.Root>
-              <Select.Trigger data-testid="trigger" />
-              <Select.Positioner />
-            </Select.Root>
-          </Field.Root>,
-        );
-
-        const trigger = screen.getByTestId('trigger');
-
-        expect(trigger).not.to.have.attribute('aria-invalid');
-
-        fireEvent.focus(trigger);
-        fireEvent.blur(trigger);
-
-        await flushMicrotasks();
-
-        expect(trigger).to.have.attribute('aria-invalid', 'true');
-      });
-    });
   });
 
   describe('prop: validationMode', () => {
